Revert "Tabs: VR implementation changes (#3763)"
This reverts commit 45ee62b263f8388c5020b1424ade5412b3100cb6.
