+++ packages/gestalt-design-tokens/tokens/vr-theme-web-mapping/comp-web-color-dark.json
@@ -307,30 +307,6 @@
             "value": "{sema.color.pressed.background.default.value}"
           }
         }
-      },
-      "tabs": {
-        "default": {
-          "base": {
-            "value": "{sema.color.background.default.value}"
-          },
-          "hover": {
-            "value": "{sema.color.hover.background.default.value}"
-          },
-          "active": {
-            "value": "{sema.color.pressed.background.default.value}"
-          }
-        },
-        "transparent": {
-          "base": {
-            "value": "{sema.color.background.default.value}"
-          },
-          "hover": {
-            "value": "{sema.color.hover.background.default.value}"
-          },
-          "active": {
-            "value": "{sema.color.pressed.background.default.value}"
-          }
-        }
       }
     },
     "border": {
