+++ packages/gestalt-design-tokens/tokens/vr-theme-web-mapping/comp-web-color-light.json
@@ -292,30 +292,6 @@
             "value": "{sema.color.pressed.background.default.value}"
           }
         }
-      },
-      "tabs": {
-        "default": {
-          "base": {
-            "value": "{base.color.transparent.value}"
-          },
-          "hover": {
-            "value": "{sema.color.hover.background.default.value}"
-          },
-          "active": {
-            "value": "{sema.color.pressed.background.default.value}"
-          }
-        },
-        "transparent": {
-          "base": {
-            "value": "{base.color.transparent.value}"
-          },
-          "hover": {
-            "value": "{sema.color.hover.background.default.value}"
-          },
-          "active": {
-            "value": "{sema.color.pressed.background.default.value}"
-          }
-        }
       }
     },
     "border": {
